Another couple of tweaks to :spectrum-generator, allowing you to limit stochastic mutation to a specific window.
Stochastic mutation is performed per-"bin" on each candidate spectrum. If a randomly-selected number is less than the given "stochastic.mutation" parameter, then mutation is performed:
 - select a random-number in [ -window/2, +window/2 ]
 - add that number to the current "bin"
 - clamp that bin to [ minEnergy, maxEnergy ]

Previously, the value for "window" was hardcoded to 0.2 -- i.e., each mutation would be confined to, at most, +/-0.1 from its previous value.

By giving the user the ability to tailor the mutation-window for each execution, the user may be able to use the stochastic search function (which appears to be much more time-efficient than brute-force) to fine-tune an existing spectrum.