Reworked GUI to use Vaadin's Navigator. Created custom Navigator implementation to hook into Security infrastructure. Enhanced RenderList buttons.
FrontEndUI now has 2 main components:
 - MainMenuBar
 - a ViewDisplay (which simply houses the current View)

Views are then selected by the Navigator, which uses the current URI to determine the active view-name.

SecurityConsciousSpringNavigator is a Navigator extension that will evaluate each navigation-selection and determine if the navigation is allowed. If the navigation is not allowed, it redirects the user to the UnauthorizedView.

Created 2 new annotations which the SecurityConsciousSpringNavigator can use:
 - AuthorizedView -- denotes which authorities a principal MUST have to navigate to a View
 - AuthorizedViews (contains multiple AuthorizedView instances)