Fixed bugs in ObjectCreator component and our use of Spring Security.
ObjectCreator -- the constructor was calling setClass(...) on itself. Moved this call to a @PostConstruct method.

Spring Security -- we were getting errors where:
 1) the user would log in
 2) this would trigger a SuccessfulLogin message
 3) when a suitable message-handler would try to get log-in details, it wouldn't find any

This problem would disappear if the user then logged-out and -in.

The problem was caused by the EventBus spawning worker-threads (because we're using the AsyncEventBus subclass), and those worker-threads wouldn't have access to the just-configured Spring SecurityContext (because that SecurityContext is stored in a ThreadLocal SecurityContextHolder).

The solution was to extend the ThreadPoolTaskExecutor we're supplying to the AsyncEventBus.
 - each submitted task is supplied with the calling thread's SecurityContext before being passed to a new Thread.

...
	@Override
	public void execute(Runnable task) {
		final var auth = SecurityContextHolder.getContext().getAuthentication;
		super.execute(() -> {
			SecurityContextHolder.getContext().setAuthentication(auth);
			task.run();
		});
	}
...