Removed gradle.properties from source-control.
The only reason this file exists is to direct Gradle to use a specific Java version.
This configuration is, however, pretty machine-specific. Therefore, it doesn't make sense to keep this in a (supposedly machine-neutral) code repository.

If you want to direct Gradle to a specific Java version, you'll have to provide your own gradle.properties from now on.