Added a first-draft of an AdaptiveSampler.
AdaptiveSampler is basically a StratifiedSampler, with the addition of 2 features:

 - it collects all estimates (in the form of XYZ triplets) for all Samples generated on the current pixel
 - it can repeat the pixel's worth of samples (minSamplesPerPixel), providing up to maxSamplesPerPixel, if the standard deviation of the gathered estimates' luminances (i.e., Y components) do not exceed a certain limit

At present, this is not a satisfactory way of doing adaptive sampling. It does not necessarily resolve visual noise in a satisfactory way. However, it *does* take a long time. It seems to mostly generate maxSamplesPerPixel, except in a few cases where it only generates minSamplesPerPixel.