Split Render configuration into RenderSetup. Child-renders now only render and return their fragment of the total Image.
RenderSetup now holds the configuration for the render as a whole:
 - sampler
 - film
 - renderer
 - camera
 - scene reference

A RenderSetup, in turn, links to one or more Renders, representing the actual execution(s) of the single RenderSetup. A RenderSetup, after all, could be executed multiple times.

Render now stores not only width and height, but also offsetX and offsetY. These new fields control the Render's offset relative to its parent.
OffsetX and offsetY are initialized during decomposition. When a RenderTask is created (to be sent to the new-task-queue), the RenderSetup's inflated Sampler and Film instances are modified by width, height, offsetX, and offsetY. (The actual RenderSetup configurations are, of course, not modified.)